refs #2016 Switch to single-module retrieval.
After a phone call with Aaron, he clarified that the ImuModules object is really not what we want.
It's only for multi-table parallel keyword searching.  Changing to ImuModule also changes the
data structure of the return, so that had a dozen knock-on effects that had to be cleaned up as well,
including using only the base table in that fetch set (which Aaron also said we should be doing).